Move cancel-visit-success up
